+++ components/LandingPage/Title.tsx
@@ -1,8 +1,8 @@
-import { useZoraCreate } from '@/providers/ZoraCreateProvider'
+import { useZoraCreateProvider } from '@/providers/ZoraCreateProvider'
 import { Input } from '../ui/Input'
 
 const Title = () => {
-  const { name, setName } = useZoraCreate()
+  const { name, setName } = useZoraCreateProvider()
 
   return <Input value={name} onChange={(e) => setName(e.target.value)} />
 }
